+++ phpmyfaq/translations/language_en.php
@@ -1487,5 +1487,7 @@ $PMF_LANG['msgPluginDescription'] = 'Description';
 // added v4.1.0-alpha.2 - 2025-02-23 by Thorsten
 $PMF_LANG['msgReportABug'] = 'Report a bug';
 $PMF_LANG['msgDashboard'] = 'Dashboard';
+$PMF_LANG['msgOrphanedFAQs'] = 'Orphaned FAQs';
+$PMF_LANG['msgOrphanedFAQsDescription'] = 'These FAQs are not assigned to any category. You can assign them to a category here.';
 
 return $PMF_LANG;