+++ phpmyfaq/translations/language_ur.php
@@ -1439,5 +1439,6 @@ $PMF_LANG['msgDocumentation'] = 'دستاویزات';
 // شامل کیا گیا v4.1.0-alpha.2 - 2025-02-23 by Thorsten
 $PMF_LANG['msgReportABug'] = 'ایک مسئلہ رپورٹ کریں';
 $PMF_LANG['msgDashboard'] = 'ڈیش بورڈ';
+$PMF_LANG['msgOrphanedFAQs'] = ' یتیم FAQs';
 
 return $PMF_LANG;